+++ src/LiveComponent/assets/src/UnsyncedInputContainer.ts
@@ -1,6 +1,6 @@
 export default class UnsyncedInputContainer {
     #mappedFields: Map<string, HTMLElement>;
-    #unmappedFields: Array<HTMLElement>= [];
+    #unmappedFields: Array<HTMLElement> = [];
 
     constructor() {
         this.#mappedFields = new Map();
@@ -18,14 +18,6 @@ export default class UnsyncedInputContainer {
 
     all() {
         return [...this.#unmappedFields, ...this.#mappedFields.values()]
-    }
-
-    clone(): UnsyncedInputContainer {
-        const container = new UnsyncedInputContainer();
-        container.#mappedFields = new Map(this.#mappedFields);
-        container.#unmappedFields = [...this.#unmappedFields];
-
-        return container;
     }
 
     allMappedFields(): Map<string, HTMLElement> {